+++ sdk/storage/azure-storage-queue/azure/storage/queue/aio/_queue_client_async.py
@@ -25,7 +25,7 @@ except ImportError:
     from urlparse import urlparse  # type: ignore
     from urllib2 import quote, unquote  # type: ignore
 
-from azure.core.exceptions import HttpResponseError, ResourceExistsError
+from azure.core.exceptions import HttpResponseError
 from azure.core.tracing.decorator import distributed_trace
 from azure.core.tracing.decorator_async import distributed_trace_async
 
@@ -153,28 +153,6 @@ class QueueClient(AsyncStorageAccountHostsMixin, QueueClientBase):
             )
         except HttpResponseError as error:
             process_storage_error(error)
-
-    @distributed_trace_async
-    async def create_if_not_exists(self, **kwargs):
-        # type: (Any) -> None
-        """Creates a new queue in the storage account.
-
-        If a queue with the same name already exists, it is not changed.
-
-        :keyword Dict(str,str) metadata:
-            A dict containing name-value pairs to associate with the queue as
-            metadata. Note that metadata names preserve the case with which they
-            were created, but are case-insensitive when set or read.
-        :keyword int timeout:
-            The server timeout, expressed in seconds.
-        :return: None or the result of cls(response)
-        :rtype: None
-        :raises: StorageErrorException
-        """
-        try:
-            return await self.create_queue(**kwargs)
-        except ResourceExistsError:
-            return None
 
     @distributed_trace_async
     async def delete_queue(self, **kwargs):
@@ -486,11 +464,13 @@ class QueueClient(AsyncStorageAccountHostsMixin, QueueClientBase):
         # type: (Optional[Any]) -> AsyncItemPaged[QueueMessage]
         """Removes one or more messages from the front of the queue.
 
-        When a message is retrieved from the queue, the response includes the message
+         When a message is retrieved from the queue, the response includes the message
         content and a pop_receipt value, which is required to delete the message.
         The message is not automatically deleted from the queue, but after it has
         been retrieved, it is not visible to other clients for the time interval
-        specified by the visibility_timeout parameter.
+        specified by the visibility_timeout parameter. The iterator will continuously
+        fetch messages until the queue is empty or max_messages is reached (if max_messages
+        is set).
 
         If the key-encryption-key or resolver field is set on the local service object, the messages will be
         decrypted before being returned.
@@ -511,6 +491,8 @@ class QueueClient(AsyncStorageAccountHostsMixin, QueueClientBase):
             should be set to a value smaller than the time-to-live value.
         :keyword int timeout:
             The server timeout, expressed in seconds.
+        :keyword int max_messages: 
+            An integer that specifies the maximum number of messages to retrieve from the queue.
         :return:
             Returns a message iterator of dict-like Message objects.
         :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.storage.queue.QueueMessage]
@@ -527,6 +509,7 @@ class QueueClient(AsyncStorageAccountHostsMixin, QueueClientBase):
         messages_per_page = kwargs.pop('messages_per_page', None)
         visibility_timeout = kwargs.pop('visibility_timeout', None)
         timeout = kwargs.pop('timeout', None)
+        max_messages = kwargs.pop('max_messages', None)
         self._config.message_decode_policy.configure(
             require_encryption=self.require_encryption,
             key_encryption_key=self.key_encryption_key,
@@ -540,7 +523,7 @@ class QueueClient(AsyncStorageAccountHostsMixin, QueueClientBase):
                 cls=self._config.message_decode_policy,
                 **kwargs
             )
-            return AsyncItemPaged(command, results_per_page=messages_per_page, page_iterator_class=MessagesPaged)
+            return AsyncItemPaged(command, results_per_page=messages_per_page, page_iterator_class=MessagesPaged, messages_to_retrieve=max_messages)
         except HttpResponseError as error:
             process_storage_error(error)