+++ sdk/storage/azure-storage-queue/tests/test_queue.py
@@ -71,28 +71,6 @@ class StorageQueueTest(StorageTestCase):
         self.assertTrue(created)
 
     @QueuePreparer()
-    def test_create_queue_if_not_exists_without_existing_container(self, storage_account_name, storage_account_key):
-        # Action
-        qsc = QueueServiceClient(self.account_url(storage_account_name, "queue"), storage_account_key)
-        queue_client = self._get_queue_reference(qsc)
-        created = queue_client.create_if_not_exists()
-
-        # Asserts
-        self.assertTrue(created)
-
-    @QueuePreparer()
-    def test_create_queue_if_not_exists_with_existing_container(self, storage_account_name, storage_account_key):
-        # Action
-        qsc = QueueServiceClient(self.account_url(storage_account_name, "queue"), storage_account_key)
-        queue_client = self._get_queue_reference(qsc)
-        queue_client.create_queue()
-        created = queue_client.create_if_not_exists()
-
-        # Asserts
-        self.assertIsNone(created)
-        
-
-    @QueuePreparer()
     def test_create_queue_fail_on_exist(self, storage_account_name, storage_account_key):
         # Action
         qsc = QueueServiceClient(self.account_url(storage_account_name, "queue"), storage_account_key)
@@ -394,6 +372,39 @@ class StorageQueueTest(StorageTestCase):
         # Asserts
         self.assertIsNotNone(result)
         self.assertEqual(4, len(result))
+
+        for message in result:
+            self.assertIsNotNone(message)
+            self.assertNotEqual('', message.id)
+            self.assertNotEqual('', message.content)
+            self.assertNotEqual('', message.pop_receipt)
+            self.assertEqual(1, message.dequeue_count)
+            self.assertNotEqual('', message.inserted_on)
+            self.assertNotEqual('', message.expires_on)
+            self.assertNotEqual('', message.next_visible_on)
+
+    @QueuePreparer()
+    def test_get_messages_with_max_messages(self, storage_account_name, storage_account_key):
+        # Action
+        qsc = QueueServiceClient(self.account_url(storage_account_name, "queue"), storage_account_key)
+        queue_client = self._get_queue_reference(qsc)
+        queue_client.create_queue()
+        queue_client.send_message(u'message1')
+        queue_client.send_message(u'message2')
+        queue_client.send_message(u'message3')
+        queue_client.send_message(u'message4')
+        queue_client.send_message(u'message5')
+        queue_client.send_message(u'message6')
+        queue_client.send_message(u'message7')
+        queue_client.send_message(u'message8')
+        queue_client.send_message(u'message9')
+        queue_client.send_message(u'message10')
+        pager = queue_client.receive_messages(max_messages=5)
+        result = list(pager)
+
+        # Asserts
+        self.assertIsNotNone(result)
+        self.assertEqual(5, len(result))
 
         for message in result:
             self.assertIsNotNone(message)