+++ sdk/storage/azure-storage-queue/CHANGELOG.md
@@ -3,7 +3,11 @@
 ## 12.2.1 (Unreleased)
 
 ### Features Added
-- Added support for `create_if_not_exists()` for `QueueClient`
+- Added support for `max_messages` in `receive_messages()` to specify the maximum number of messages to receive from the queue.
+
+### Other Changes
+- Updated documentation for `receive_messages()` to explain iterator behavior and life-cycle.
+- Added a sample to `queue_samples_message.py` (and async-equivalent) showcasing the use of `max_messages` in `receieve_messages()`. 
 
 ## 12.2.0 (2022-03-08)